rkowlu/DeepLabV3Plus-Pytorch/requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
tqdm
matplotlib
visdom
# Optional for Road Anomaly sanity check: pip install kagglehub
kagglehub